+++ R/periods.R
@@ -3,7 +3,7 @@
 lubridate_period <- function(...){
   periods <- unclass(period_timespan(...))
   # lubridate::period() propagates NA values across all periods
-  if (cheapr::num_na(periods, recursive = TRUE) > 0){
+  if (cheapr::na_count(periods, recursive = TRUE) > 0){
     which_na_fill <- which(cheapr::row_any_na(list_as_df(periods)))
     for (i in seq_along(periods)){
       periods[[i]][which_na_fill] <- NA