+++ NEWS.md
@@ -1,4 +1,47 @@
-# timeplyr 0.9.0
+# timeplyr (Development version)
+
+### Major breaking changes
+
+- Time intervals in timeplyr have been re-imagined.
+They are now fixed-width intervals which means any one time interval vector
+will have a common width for all intervals it contains, e.g. 1 month.
+This makes internal operations much faster and simplifies the logic
+for working with these objects. The data structure for these intervals
+is a vector of LHS start times with an attribute to specify the
+timespan or width of the interval. All intervals remain LHS closed and
+RHS open. As a side-note this new data structure makes it possible for 
+time intervals to be used in `data.table`.
+
+- Most functions which are not time related have been removed.
+
+- Many arguments have been renamed or removed. 
+  - This includes the `time_type`
+  argument which is no longer used. When specifying units such as 
+  days, weeks, months or years, exact timespans will be used. 
+  - Most functions no longer use the args `roll_dst` and `roll_month` with the 
+  exception of the sequence functions.
+  - The `time_by` argument has been renamed to either `width` or `timespan`
+  depending on the context.
+  
+- Functions `time_expandv`, `time_completev`, `time_span_size` have been 
+renamed to `time_grid`, `time_complete_missing`, `time_grid_size` respectively.
+
+- Functions `time_summarisev` and `time_aggregate` have been removed, use 
+`time_cut_width` instead.
+
+- `time_cut` has been deprecated and renamed to `time_cut_n`.
+
+- `edf`, `asc` and `desc` have been removed as they do not fit the 
+purpose of the package.
+  
+### New features
+
+- A new custom timeplyr object `timespan` to create and use 
+timespans for various time operations.
+
+- New functions `resolution` and `granularity` to calculate 
+the resolution and granularity of a time vector respectively.
+See the help page `?resolution` for more details.
 
 ### Upcoming breaking changes